requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
jinja2==3.1.2
pydantic==2.4.2 